fix: hsl promote to double. Hue is a colour space, not a set enum, like in RGB.

+++ inc/gfx_color.h
@@ -6,7 +6,7 @@
 /*   By: nmartins <nmartins@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/04/30 00:56:50 by nmartins       #+#    #+#                */
-/*   Updated: 2019/05/02 20:29:43 by nmartins      ########   odam.nl         */
+/*   Updated: 2019/06/24 23:16:31 by nmartins      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct	s_rgb
 
 typedef struct	s_hsl
 {
-	int		h;
+	double	h;
 	double	s;
 	double	l;
 }				t_hsl;
@@ -43,7 +43,7 @@ int				gfx_color(int r, int g, int b, int a);
 t_rgb			gfx_hsl2rgb(t_hsl hsl);
 t_hsl			gfx_rgb2hsl(t_rgb rgb);
 t_rgb			mk_rgb(int r, int g, int b);
-t_hsl			mk_hsl(int h, double s, double l);
+t_hsl			mk_hsl(double h, double s, double l);
 
 int				gfx_color_from_rgb(t_rgb rgb);
